🚚 Move streamlined header file
