+++ static/app/views/issueDetails/streamline/header.tsx
